Исправил недочеты

+++ exercise2.rb
@@ -1,6 +1,6 @@
 puts "Введите высоту треугольника"
-h = gets.chomp.to_i
+h = gets.chomp.to_f
 puts "Введите основание треугольника"
-a = gets.chomp.to_i
+a = gets.chomp.to_f
 square = 0.5 * a * h
 puts "Площадь треугольника с основанием #{a} и высотой #{h} равна #{square}"
